Created the following new classes as per tickets: Actor, Renderer and World
Created the following classes to demonstrate functionality:
BoxActor - represents a box (derives from Actor)
Quad - represents a quad that can be rendered

Currently, the application launches to display a cube like before, but this time it's using our classes so as to demonstrate their usage.  The functionality for rotation is not yet implemented.

[git-p4: depot-paths = "//depot/": change = 157]